Write configurations to _deploy.yml or DESCRIPTION

deployr execution can be controlled through the use of a _deploy.yml file in the projects root or by setting the same values in the Config/deploy field of the projects DESCRIPTION file. The priority of settings are _deploy.yml > DESCRIPTION > defaults. config_set() allow you to write configuration values to either of these places. config_get retrieves the configuration for the current project (by default merging all sources together).

config_set(..., .in = c("deploy", "desc"))

config_get(.in = c("default", "desc", "deploy"))

Arguments

...

Named values to write to the configuration

.in

The location to write the configuration to or read it from. For config_get() multiple values are allowed and the values from the sources will be merged together according to the priority in the description.

Value

This function is called for its side effect

Details

Recognised settings

The following list provides an overview of the settings that are currently used by deployr:

distDir: The location of files generated by static projects. Defaults to ./dist/.

port: The port to use for dynamic projects. Defaults to 50505.

host: The host to use for dynamic projects. Defaults to "127.0.0.1".

workerId: An ID that identifies the specific worker running the code. Will be used by Shiny runtimes to correctly route clients to the same worker. Defaults to "".

sharedSecret: A shared secret used by Plumber, plumber2 and Shiny as a simple authentication mechanism. Defaults to NULL.

logger: The name of the R package used for logging by the server. Currently unused. Defaults to "logger".

staticFormat: The format to render to for static projects. Is passed to output_format in quarto and rmarkdown rendering pipelines. Defaults to NULL

params: A list of parameters to overwrite in the YAML front-matter of quarto and rmarkdown documents. Will be passed to quarto::quarto_render(execute_params) and rmarkdown::render(params). Defaults to list()

quartoArgs: Additional arguments for quarto, passed to quarto::quarto_render(quarto_args). Defaults to NULL

pandocArgs: Additional arguments for quarto, passed to quarto::quarto_render(pandoc_args). Defaults to NULL

hugoArgs: Additional arguments for blogdown, passed to blogdown::build_site(hugo_args). Defaults to NULL