Execute a producethis project

This function assumes the current working directory is the root of a project created according to producethis conventions. If this is the case it will execute the project according to its type and the content in ./exec

execute()

Value

This function is called for its side effects

Details

Project categories

producethis/deployr recognises three main project types:

static: The project is executed and as a result produces one or more files which will subsequently be made availabe

dynamic: The project is executed and as a result starts a web server with dynamic content

batch: The project is executed and is afterwards considered "finished". If information is generated it needs to be stored somewhere else.

Inside these categories are a range of types that maps to specific runtimes or execution engines. The project type is given in the Type field of the DESCRIPTION file

static

quarto: Used for a single quarto document or a quarto project. Will use quarto::quarto_render() to render the content of exec/

rmarkdown: Used for a single RMarkdown document. Will use rmarkdown::render() to render the .Rmd file in exec/

book Used for a bookdown project. Will use bookdown::render_book() to render the content of exec/

blog Used for a blogdown project. Will use blogdown::build_site() to render the content of exec/

site Used for general Rmarkdown websites. Will use rmarkdown::render_site() to render the content of exec/

dynamic

shiny: Used for a shiny app. Will use shiny::shinyAppDir() to create a shiny app from exec/ and then launch it with shiny::runApp()

plumber: Used for a plumber API. Will use plumber::plumb() to create an api from exec/ and launch it with plumber::pr_run()

api/app: Used for a generalized app or api. Will require a _server.yml file to specify how to launch it

rmd_shiny: Used for a shiny-backed Rmarkdown document. Will use rmarkdown::run() to render and serve a dynamic document

quarto_shiny: Used for a shiny-backed Quarto document. Will use quarto::quarto_serve() to render and serve a dynamic document.

batch

batch: Batch project which will run each script in exec/ in turn and then quit.